requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
django-crispy-forms==1.8.1
idna==2.8
mysqlclient==1.4.6
oauthlib==3.1.0
Pillow==7.0.0
pkg-resources==0.0.0
python-http-client==3.2.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
sendgrid==6.1.0
sqlparse==0.3.0
urllib3==1.25.7
XlsxWriter==1.2.7
